% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splittaxa.R
\name{split_taxa}
\alias{split_taxa}
\title{Split the taxonomic assinment into 7 taxa: Domain, Phylum, Class, Order, Family and Specie}
\usage{
split_taxa(taxonomy)
}
\arguments{
\item{taxonomy}{The taxonomy object produced by tagmeFromFasta or tagmeFromDada2 functions.}
}
\value{
A data frame with 7 columns: Domain, Phylum, Class, Order, Family and Specie, describing the taxonmy for each sequence. This data frame can be used as input for phyloseq::tax_table().
}
\description{
Split the taxonomic assinment into 7 taxa: Domain, Phylum, Class, Order, Family and Specie
}
