% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizebylevel.R
\name{summarize.by.level}
\alias{summarize.by.level}
\title{Sum the sequences that belong to the same taxon.}
\usage{
summarize.by.level(count.table, taxonomy, level = "Genus",
  taxa.are.rows = FALSE)
}
\arguments{
\item{count.table}{The count table containing the quantification for each sequence.}

\item{taxonomy}{The taxonomic assignment produced by TAG.ME.}

\item{level}{The The taxon level to be summarized. Values are: "Domain", "Phylum", "Class", "Order", "Family", "Genus", or "Specie". Default value = "Genus". Default is FALSE.}

\item{taxa.are.rows}{Boolean value describing your count table organization. TRUE value means that your table contains your sequences (ASVs, OTUs) as rownames, and your samples are colnames - Ex: OTU tables generated by Qiime. FALSE value means that your sequences are the colnames and samples are the rownames - Ex: Seqtable produced by Dada2.}
}
\value{
A count table with the unique taxons as rownames and the summarized count for each.
}
\description{
Sum the sequences that belong to the same taxon.
}
