% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmeFromDada2.R
\name{tagmeFromDada2}
\alias{tagmeFromDada2}
\title{Assign taxonomy to an input Seqtab variable, exported from Dada2}
\usage{
tagmeFromDada2(seqtab, db = "./", specificity = 0.8, batch = 50000)
}
\arguments{
\item{seqtab}{The seqtab object produced by Dada2.}

\item{db}{The directory containing the downloaded RDS and TXT files corresponding to your amplicon region. The directory name must end with "/".}

\item{specificity}{The Specificity that will determine the threshold score for assignment.}

\item{batch}{The maximum number of sequences to be assigned per batch. Bigger batches demand more memory.}
}
\value{
A table containing the taxonomic assignment and the score to each sequence.
}
\description{
Assign taxonomy to an input Seqtab variable, exported from Dada2
}
