% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmeFromFasta.R
\name{tagmeFromFasta}
\alias{tagmeFromFasta}
\title{Assign taxonomy to an input FASTA file}
\usage{
tagmeFromFasta(file, db = "./", specificity = 0.8, batch = 50000)
}
\arguments{
\item{file}{The FASTA file to be classified. It can be a group of ASV, OTU or Unique sequences.}

\item{db}{The directory containing the downloaded RDS and TXT files corresponding to your amplicon region. The directory name must end with "/".}

\item{specificity}{The Specificity that will determine the threshold score for assignment.}

\item{batch}{The maximum number of sequences to be assigned per batch. Bigger batches demand more memory.}
}
\value{
A table containing the taxonomic assignment and the score to each sequence.
}
\description{
Assign taxonomy to an input FASTA file
}
